+++ spec/components/search_spec.rb
@@ -1,0 +1,45 @@
+require "rails_helper"
+
+describe "Search", type: :view do
+  def component_name
+    "search"
+  end
+
+  it "renders a search box with default options" do
+    render_component({})
+    assert_select ".gem-c-search.gem-c-search--on-white"
+  end
+
+  it "renders a search box for a dark background" do
+    render_component(on_govuk_blue: true)
+    assert_select ".gem-c-search.gem-c-search--on-govuk-blue"
+  end
+
+  it "renders a search box with a custom label text" do
+    render_component(label_text: "This is my new label")
+    assert_select ".gem-c-search .gem-c-search__label", text: "This is my new label"
+  end
+
+  it "renders a search box with a custom label content" do
+    render_component(inline_label: false, label_text: "<h1>This is a heading 1</h1>")
+    assert_select ".gem-c-search .gem-c-search__label h1", text: "This is a heading 1"
+    assert_select ".gem-c-search.gem-c-search--separate-label"
+  end
+
+  it "renders a search box with a value" do
+    render_component(value: "I searched for this")
+    assert_select ".gem-c-search .gem-c-search__input" do
+      assert_select "[value=?]", "I searched for this"
+    end
+  end
+
+  it "renders a search box with a custom id" do
+    render_component(id: "my-unique-id")
+    assert_select ".gem-c-search #my-unique-id.gem-c-search__input"
+  end
+
+  it "renders a large search box" do
+    render_component(size: "large")
+    assert_select ".gem-c-search.gem-c-search--large"
+  end
+end